Correct the ring buffer tests
Modifies the argument order for type arguments to deque and enqueue as
well, which should be a huge deal. In managed we also expose and
modify the From converters to also have Vec in tests and expose a
constructor for a single `T` to facilitate `Default`. An actual
`Default` is not exposed for lack of clear definition, could be an empty
slice as well.