Sketch the ethernet endpoint and receiver
Added a few ?Sized bounds to the base traits from the nic since those
were missing but unecessary due to references everywhere. Note that
`eth::Frame` itself takes a value and no references to make it
potentially feasible to instantiate it with an owning buffer such as
`Vec<_>` or the libraries own `crate::managed::Slice`.