Reduce pretty printer complexity
There is no real use in providing a *reference to* a dynamic trait
object for `AsRef<[u8]>` within the pretty printer trait. May want to
change this later to a generic method (`impl Payload` or the standard
`impl AsRef<[u8]> + ?Sized`) later but this is good enough for now. The
only real use was instantiating it with an effective byte slice anyways,
or the pretty printers itself did this immediately.