Rework lifetimes bounds of devices
A receiver/sender on a nic needs to be ready to accept arbitrarily short
lifetimes on the data give to them. That is, the device is allowed to
fully control this allocation. It is not bound to at least the duration
of the reference to the device and does not borrow from it. Instead the
device feeds to packets to the receiver. This makes it possible to
create temporary state within the method and after having pushed the
data read again even from mutable state borrowed to the packet.

This is not yet complete, as the current formulation would require
higher-kinded traits to give a `Device` trait again. To remedy this, we
would need to get rid of the lifetime parameter in the `Packet` trait so
that it can become a bound on type parameters without requiring its own
lifetime parameter.