Get rid of Packet trait while reintroducing Device
Both the Handle and Payload are now associated types on the Device
instead. A standard `Packet` struct fulfills a role that mostly groups
the parameters.

Even unsized types refer, in current Rust, to contiguous memory regions.
Thus, the two references contained in the Packet struct, each referring
to one of the associated types, splits the interface for a device neatly
into two concerns: the payload to describe the block owned by the
device; and the handle describing functionality added by the driver
interface on top.