Ensure consistent naming and Payload impl on eth
What was previously named `payload` and `payload_mut` is now migrated to
instead be named `payload_slice` and `payload_mut_slice`. In its place,
the internal Payload+PayloadMut interface is put. Even though the Frame
is not DerefMut for `ethernet`, the payload is the exception that can be
mutably accessed without breaking consistency between header and repr.