Provide type erased nic handle to eth layer
Type erasure provides us with a remedy against potential growth of type
parameters, and will also scale to deeper layers. Upper layers will no
longer make use of the direct frame api but instead try to get an
ethernet buffer, then an ip buffer and so on. The exact set of traits
and their functions is subject to change.