Replace cfg based addition of managed Vec variant
The additional existance of that variant made it fairly ugly to handle
its existance at every callsite (since existance is a public interface).
Thus, we replace `Vec` with an uninhabited type in `no_std` environment.
While the exact interface is of course still dependent on the
configuration, the relevant `&self`-based interface is replicated
without actual implementation. This remove the necessity to `cfg()` the
variants of some matches (and should not lead to any actual overhead).