Simplify ethernet module
Removes the abstract trait for the ethernet layer. We commit to a single
implementation whichever functionality it will have in the end. This
does not exclude that one layer may be able to use a combination of
layers below. Simplified receivers that are oblivious of the layer of
which they receive payload are also not blocked by this.

It greatly simplifies implementing layers on top though by knowing the
precise methods available for routing and writing an ethernet frame.

The payload itself is still behind a trait. This must stay a type
parameter in order to allow mutable and immutable packet types to be
received but on mutable ones sent.