Push timekeeping to the nic, basic ipv4 layer
Simplifies tests and makes clock resolution to packet accuracy
potentially feasible. It is expected that nic that cares about that will
recalculate the timestamp when queueing a packet for example to assure
its accuracy for loopback causes. The `Info` struct is explicitely
allowed to change on calling any of the mutable methods of `eth::Handle`
to enable such behaviour.

Haven't figured out the precise behaviour for ipv6 fragmented packets
yet or how to calculate checksums after having sent the packet. But that
should be solvable by reborrowing the payload with unchecked repr in the
sender case. This also provides the opportunity to do some last-minute
fixup.