Add udp endpoint basics
This is unlikely the final version. Giving the source addresses and port
explicitely for every packet seems rather inconvenient but was
inherited from the ip endpoint design. Since this requires a lookup for
every sent packet (finding one fitting source address) it is also
overhead we want to avoid. An alternative is providing the address as an
index based lookup instead or as a variant.

This design will come into question once more with TCP where connection
state needs to be present for (src, src_port, dst, dst_port) tuples.