Sketch icmp layer
Does not yet include a good story for handling unknown icmp types
(technically should forward them to an upper layer if available). This
layer is also the first to need good integration for rewriting incoming
packets.

The next step is to introduce reinintialization of packets so that we
can answer echo requests zero-copy. The plan is thus to split the
current packet representation in three (In, Raw, Out) and to introduce
`reinit` methods on the layers below. More on the design internals
follows since zero-copy is tricky when we have no memory management of
our own.