Eth layer reinitialization
Creates an out of zero-copy retransmission of a packet. In particular
the payload needs not be touched if the packet can be changed inline
that is all headers of all layers have the same size. Even if not we
touch less cache lines in doing so. Work in progress, untested.