Utilize ip reinitialization for icmp
This should work fine since the resulting packet always has the smallest
possible representation, never exhausting the available buffer space. It
also ensures to never copy the buffer if possible to avoid since only
the type in the icmp header needs to be rewritten (the implementation
does rewrite the other header fields as well for now but not the body).