Mark out packets as must_use and send icmp answer
Previously forgot to actually commit the constructed answer to be sent.
This makes it reasonable to mark all packets as must_use to make this
mistake less likely. It is still not perfect as you can fill in the
payload slice (a use) and then forget to send the packet without any
warning but it's a start. Cleans up some unused ip reinit parts