Provide str parsing for some addresses
The parsing of ipv4 cidr is only available with `std` enabled as it will
use the std provided parsing for the address part. This avoid recoding
the relevant RFC for now but should be addressed in the future. Ethernet
addresses can however be parsed always and the standard library would
also offer no additional help.