Ip cidr: Split subnet and address representation
The wire layer no longer utilizes the broadcast and network identifying
addresses as a in-band representation of the whole subnet block. This
clarifies a few code paths in particular for IPv6 where no such
addresses exist. The ip endpoint will now always check that the
addresses given to it are unicast to avoid misuse.

This enables route selection for ipv6 packets possible. Note that
incoming packets are not yet recognized as belonging to the endpoint.