Cleanup of example and sys interface
Remove two calls to `dbg!` that were remnants from the time when the tap
interface did not yet appear to work. Polished the output of the example
to show when a packet was answered.

Remove the `arp` layer replacement as it had no effect. The linux kernel
apparently does not accept such an announcement by default and thus the
guide for the example simply requires a static arp entry on the host
which should work across more systems anyways. This also slims down the
perceived complexity of the example considerably.