Sanity for tcp
Remove previously messed up entangling of checksum argument
dependencies, interleaving of parsing and Packet construction.
Introduction of an unsized newtype for the underlying packet bytes will
follow at a later point to eliminate the current requirement of a dummy
representation for parsing and emitting that is internally created and
unused.