Avoid malformed crash on unknown flag combinations
Since this has no real influence on the packet structure (individual
flags that have can still be tested independent from this) it should be
handled by the layer and not the packet representation. The RFC is also
not clear whether packets with the unhandled combinations need to be
handled or not but the imperative style makes it seems like some flags
remain simply unhandled when an RST is included.

This also syntactically and semantically fixes the other tests but does
not change their behaviour.