Abstract entry as basic generic map primitive
We can build get_mut and insert on-top of entry. Since there are both
owning and non-owning, Copy and non-Copy version we have to take some
care to allow insertion and removal which may require cloning in one of
those cases. Creating an Entry-api makes it possible to split these into
another enum if required.