Expand Map to a insertion and removal
This commits to a 'basic' set of operations and doesn't yet optimize for
any fancy one that would allow in-place modification for example.
Basically all of std::BtreeMap but avoiding having to move arguments and
return values instead allowing their ownership to be retained by the map
(to keep an initialized version).