Add a managed map
Intended to offer the basic functionality of a BTreeMap where available
and a replacement where it is not. For extremely resource constrained
or simple environments with constant size a slice based version with
simple linear search and Stack insertion should also be available.
Actual accessor functionality to be added later.