Use SipHash for isn generation
Assuming some initial randomness this should be more than good enough to
generate very good, and unpredictable isns. Each round has ~14 primitive
operations for a total of ~150 integer computations per hashing for isn.
A few of the instructions can be interleaved but we'll likely not be
able to *accept* connections at line rate. Duh. That was not the goal
anyways.