Provide connection four-tuple key modification
This is then used immediately in the handling of an incoming SYN for a
listening connection. It may not be as nice as possible due to a few
indirections involved but overall it is already very positive to have it
working and borrow checked. The Entry interface for a connection also
provides an interface to the isn generator that utilizes the set four
tuple.