Dispatching logic to packet answer variants
Prepares the packet buffer for the operations to follow. Each variant
and accompanying state describe possible operations that can be executed
on the packet buffer. The main attributes are whether there is data to
be receive or not (may be empty), whether the buffer is free to use or
holds necessary control messages, and whether the connection is still
alive.