Start a socket-y client interface
It's not really a socket but you'll likely not want to touch the direct
implementation of `tcp::Recv`/`tcp::Send`. This one will drop packet for
all other connections which is fine for example and very simple cases
where no other connections exist. It does also not corrupt the other
state since these should retransmit/restart if they are called again
(although they may of course have timed out).

A full-stack solution managing multiple clients, server, .. is probably
not going to get added soon but it specific implementations can always
be added by the library user as well (the Recv trait is public) and the
interface tries to at least help you out. It is only necessary to route
receive requests to the correct `Client` by using the `SlotKey`, then
forward the packet to the impl on `Client`.