Send ACK to remote fin
Note to self: Don't set FIN on that ack. That can thoroughly confuse the
Linux network stack and it might ignore the first byte of the next
connection if it comes from the exact same combination of ports. Weird
stuff to get an 'method ET is not support from the web server'. It could
also be connected to the half-open connection which is also somewhat
irregular (but perfect valid and potentially less resource intensive).