Throw tcp_hello into a mininet script
Execute in the directory of the script. Opens the iface as a raw socket,
instead of a tap interface as previously, as that is how mininet exposes
the adapter. Also resets the connection when it receives the SYN+ACK
back for some currently unknown reasons.

Not perfect yet but a sketch of functionality and also potentially
automated tests. We 'curl' the expected answer prior to running the
ethox binary so that in the future (when we also store the actual result
as well) we can compare them. Magic.