An unsafe handle wrapper and nic loss adapter
This, sadly, requires unsafe to implement the wrapper without running
into lifetime problems from the associated type required for a
nic::Device. This may be fixed in the future when GADT allow us to
specify the Payload and Handle as parametric over a lifetime themselves
instead of having them behind a reference in Packet to crate that
parametricity.