Fix std::time::Instant to our Instant conversion
The internal implementation of the standard library has a `zero`
timestamp but never exposes it. So we need to construct our own
reference point which will, on systems that can not guarantee a strictly
monotically increasing clock by default, now invoke two (spin-)locks on
the path to constructing the current timestamp. Nice.