Mininet: Wait, timeout and disable host tcp
We ensure that the host does not mistakenly try to answer tcp layer for
us by disabling ip addresses. This would usually also require us to do
arp but since that was performed at latest when we invoked curl just
before we are safe.

Also timeout the tcp_hello just in case.