Plan: be more principled with features
Dependency on `std` is not necessary to offer libc based sockets. On the
contrary, one popular way to create functional binaries of smaller size
is a dynamically linked libc without the standard library.

Also, `alloc` and `std` are no longer the same since a few releases ago.
We can have `Vec` and other allocated datastructures without all other
OS abstractions.

This is part of a patch series to rework the features flags to match the
semantics more precisely.