Rework sys interface to direct libc errno
Instead of allocating/converting to the std::io::Error instantly we work
with the raw errno internally. This makes it possible to avoid a
dependency on `std::io` for good.

Part of a patch series for remodelling feature flags. There is still a
dependency on SystemTime so `sys` only does not work yet.