Add a lifetime parameter to lossy
This is the same design as all other layers. Some persistent state is
kept and referenced inside the actual implementation. This could be
slightly more generic by introducing a type parameter for the kind of
borrow of the loss but that will be more relevant if we intend to add
more loss models instead.