Use a larger source with repeats
While surprising that llvm does not pick up on the pattern and does this
by itself, this increases throughput of the pattern fill operation by
8-10 times. The benchmarks are inredibly stable in a single allocation
but deviate between. That is very likely related to exact layout, which
may change between runs, such as whether the output is allocated in a
single page/cache line or spans multiple pages. Improving the
consistency by covering more layouts may be helpful for testing more
initialization techniques.