Improve UX of iperf binary
Results are not complete but clearly indicate which parts are missing.
Some remaining crashes when unexpected TCP packets arrive too late may
still exist but are not triggered regularly on a well behaved remote.

Note that this switched to a raw socket instead of a tap interface. This
makes it possible to use on a real network setup.