Rough draft for udp iperf server
We need this since the Linux version can not keep up with our packet
rates once we get to a couple of Mpps. The original iperf2 apparently
needs a syscall for each incoming packet which is definitely too slow
and many packets are already dropped in the kernel. If a mmsg based
version could keep up is unclear but at this point we may as well
rewrite it in Rust.