Fix missing end marker
When the total sent bytes were divisible by the packet size then the end
marker would be missing. This was due to a botched comparison where the
code responsible for setting the bit did not realize it was initializing
the last packet, i.e. remaining would be 0 after that packet.