Fixes loss figure
Somewhat awkward but the server side calculation of packet loss was
broken. Here's a postmortem on impact and fixing old data.

The client side of iperf can accurately count all packets it sends but
it can not know how many arrive. The server side on the other hand can
count the arriving ones but does not know how many are intended. Thus,
the client embedds a 31-bit sequence number in a 32-bit field. The last
remaining bit serves as an 'end-of-stream' flag.

The server side then gets the intended number of packets on the stream
as the maximum of sent sequence numbers (which assumes < 2**31 packets
in total but that's fine for measurements in a timespan of seconds).
However, it failed to delete the flag in the highest bit. The number
`max_received_id` is thus 2**31 bit larger than in reality. From this,
the `total_count` of packets is derived. An off-by-one error happened
here as well, the `id` is an index but a count was intended.

Next, the lost packets is calculated as:

> ((self.total_count - self.packet_count) as f32)/(self.total_count as f32)

Since `total_count` is too large, this figure tends to 1 as the packet
sizes get larger, and the count goes to 0. It is rather useless.

Lucikyl the exact calculated numbers were also printed, allowing us to
derive actual numbers without the errors made above.

> [+] Done
> Sending server report
> [..] 0.000 ms <A:total_count - packet_count>/ <B:packet_count> (<loss>%)

Thus:
actual_total_count = (A + B + 1) - (1 << 31);
actual_lost_count = (A + 1) - (1 << 31);
actual_loss = actual_lost_count / actual_total_count