Add a simple bpf module with jit
Thankfully most of the work on jit and bpf has already been open sourced
by NetBSD and been packaged into a crate. This makes the adaption code
fairly small, basically only requiring a stable interface wrapper
around and an adaptor for ethox.

The only trouble is that the crate is not `no_std` but so is neither of
the alternatives. The NetBSD code supposedly works on a larger number of
targets while other crates based on uBPF (e.g. `rbpf` crate) only JIT
for x86 and somehow even mess that up on Windows. The implementation
also doesn't detect its host so that these are incredibly unsafe.

Some part of that may still be useful later since a VM might be a
low-performance alternative where no JIT is available.