Add the icmp documentation
This also adds a bunch of basic new methods that expose the internal
configuration of the layer. This enables silencing responses and the
basic logic for an upper layer handling. A part of the documentation,
besides common parts present in almost all layers, is geared towards
this explaining this as it is a difference to other layers and vital to
the message of customizability.