Add ethernet layer documentation
This corrects some minor irregularities that were leftover from this
being one of the earliest modules written. In particular, there was no
forwarding impl of the Recv and Send traits for mutable references
unlike in the ip layer and a method to access an incoming frame that was
superseeded by the member being public, it was only being used in a test.

The re-exports of ARP were also removed as they might have been
contradictory to the appearance of a very simple layer.