Simplify pretty print
This adds a (defaulted) function to trait but is still a simplification
since it removes shared state. The indent argument is no longer mutable
but provided by-value, the formatters are more flexible with being
provided with a parsed frame or unparsed buffer, and ip/udp/tcp pretty
printing of checksums is finally shared.