Complete the sys nic documentation
For some reasons, glob exports can lead to incorrect unreachable pub
diagnostics for the items within the glob if the path preceding it is
not public, i.e. they are not simply re-exports.  Fix this by renaming
the sys module internally and exporting the export section instead as a
public version of this module.

This then leads to several detections of actual pub constants, traits,
and members that are inaccessible. They are easily fixed by using
pub(crate) instead or removing the pub attribute entirely.