Allow the unused endpoint attribute
All other TCP message require the attribute to access the connection. In
case a utility method for opening, closing, or otherwise modifying the
original or other connections are to be added the Closing will also
require access to the endpoint. Not removing it for the moment ensures
that we do not accidentally take ownership in other code and that it is
possible to extend it that way.