Disable `no_std` for doctests
The reason for this is twofold:

Firstly, the `std` feature is enabled by default and this makes writing
documentation more consistent with that fact.

Secondly, and more importantly, the standard library doctest runner
seems to both not link with an allocator (unlike the test runner) but at
the same time for some reason require it.