Merge documentation pages within chromium
This works more reliable than trying to get both chromium printer and
the merging program to work out the margins and remove whitespace from
pages. If this ever becomes a problem due to space usage (from the large
dom) then we can always move to merging all first-level modules within
the browser and the rest within the pdf files.