Fix and sync slot map with better tests
This fixes a bug that occurs when inserting elements after having
removed more than one element. While advancing the free list, a
comparison that checked if the index was wrapping or not had the wrong
sign.

When the index wrapped then the computed index would not wrap and point
to an index past the capacity limit. And when the index would not have
wrapped the capacity was wrongly subtracted for what was assumed to be
compensation for wrapping. This would lead to a negative index if the
next free list element was not the immediate successor which leads to a
very large index.

Both of these issues could not make the free list point into the valid
element list, essentially forgetting the elements and reducing the
available space.