Try to get an RX timestamp from software
It's still unclear what to do about send timestamps. As far as I've
understood one course is to use an ioctl but that is a syscall and
somewhat awkward. Wouldn't it be great if those were supported in
io-uring as well and we could just occasionally insert a query for the
current time? That'd be pretty neat. But ioctl are far away right now
and their baseline support might not even appear in 5.6.